# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFlow Runtime API
  description: |
    OpenAPI contract for endpoints currently exposed by `cmd/agentflow` runtime.

    Notes:
    - API key authentication is accepted via `X-API-Key` header.
    - Query-string `api_key` is disabled by default for security reasons.
  version: 1.3.0
  contact:
    name: AgentFlow Team
    url: https://github.com/BaSui01/agentflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: HTTP API server
...
paths:
...
  /api/v1/config:
    get:
      tags: [Config]
      summary: Get current runtime config (sanitized)
      operationId: getConfig
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Current sanitized configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
    put:
      tags: [Config]
      summary: Update runtime config fields
      operationId: updateConfig
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '200':
          description: Update applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '400':
          description: Invalid update request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
  /api/v1/config/reload:
    post:
      tags: [Config]
      summary: Reload config from file
      operationId: reloadConfig
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Reload succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '500':
          description: Reload failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
  /api/v1/config/fields:
    get:
      tags: [Config]
      summary: List hot-reloadable fields
      operationId: listConfigFields
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Hot-reloadable fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
  /api/v1/config/changes:
    get:
      tags: [Config]
      summary: List recent config changes
      operationId: listConfigChanges
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        '200':
          description: Change log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req-abc123"
...
    ConfigUpdateRequest:
      type: object
      required: [updates]
      properties:
        updates:
          type: object
          additionalProperties: true
    ConfigResponse:
      description: Config API responses use the unified Envelope structure. The data field contains config-specific fields.
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                config:
                  type: object
                  additionalProperties: true
                fields:
                  type: object
                  additionalProperties: true
                changes:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                requires_restart:
                  type: boolean
...